requirements.txt
livekit-agents
firebase-admin
flask
python-dotenv
openai-whisper
sounddevice
scipy
gTTS
llama_cpp_python